requirements.txt
streamlit
python-dotenv
langchain-google-genai
langchain-core
pyperclip